% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\docType{data}
\name{elements}
\alias{elements}
\title{Data from the periodic table of elements}
\format{A data frame with 118 rows and 3 variables:
\describe{
  \item{Z}{Atomic number}
  \item{Symbol}{Chemical symbol}
  \item{Name}{Name of element}
}}
\source{
\url{https://en.wikipedia.org/wiki/Symbol_(chemistry)}
}
\usage{
elements
}
\description{
A dataset containing atomic numbers, chemical symbols, and
names of 118 elements.
}
\keyword{datasets}
