% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_us_states.R
\docType{data}
\name{US_states}
\alias{US_states}
\title{US state names and abbreviations}
\format{A data frame with 51 rows and 10 variables:
\describe{
  \item{Name}{US Region Name}
  \item{Type}{State, State (Commonwealth), or Federal District}
  \item{Capital}{State Capital}
  \item{ANSI}{American National Standards Institute 2-letter code}
  \item{ANSI_digit}{American National Standards Institute 2-digit code. This is stored as a character to ensure all entries are two characters long}
  \item{ISO}{International Organization for Standardization 3166-2 code for US States}
  \item{USPS}{United States Postal Service 2-letter code}
  \item{USCG}{United States Coast Guard 2-letter code (used as prefices for vessel numbers)}
  \item{GPO}{US Government Printing Office abbreviations}
  \item{AP}{Associated Press Stylebook abbreviations}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_abbreviations}
}
\usage{
US_states
}
\description{
A dataset containing names of US states and their abbreviations under various standards
including ANSI, ISO, USPS, USCG, GPO, Associated Press.
}
\keyword{datasets}
